requirements.txt
Django==1.10.2
djangorestframework==3.4.7
mysqlclient==1.3.9
psycopg2==2.6.2
